requirements.txt
streamlit
pypdf
python-docx
python-pptx
pandas
langchain
faiss-cpu
google-generativeai
pydantic
python-dotenv
lark
sentence-transformers